Added a phrase to the rental contract about $9/person/day for lunch. Rentals that use the Seminar House must house at least 14 people - it used to be 15.
A green scene donation timestamp will be the same as the registration
deposit timestamp instead of being 'now'.

You can now change a program from an extra day program to a 'normal' program
by changing the end date and the 'extra days' field to 0 - and the full tuition to 0.  There are two preconditions for being able to do this:

1 - There cannot be any meeting space bookings.
2 - There cannot be any registrations that come early or leave late (i.e. before or after the program dates).